Buffer Override: fix lack of pixel quantization due to implicit floating point to integer conversion (as well as a few more implicit conversions)

+++ dfxgui/dfxguislider.cpp
@@ -827,7 +827,7 @@ DGAnimation::DGAnimation(DfxGuiEditor*		inOwnerEditor,
 						 DGImage*			inAnimationImage, 
 						 size_t				inNumAnimationFrames)
 :	DGControl<VSTGUI::CAnimKnob>(inRegion, inOwnerEditor, dfx::ParameterID_ToVST(inParameterID), 
-								 inNumAnimationFrames, inRegion.getHeight(), inAnimationImage)
+								 static_cast<int32_t>(inNumAnimationFrames), inRegion.getHeight(), inAnimationImage)
 {
 	assert(inNumAnimationFrames > 0);
